add document type for handling the requests it's work but its not effecient and will cause bad code so i try to change it

+++ src/GalaxusIntegration.Application/Commands/Products/UpdateProductPriceCommand.cs
@@ -5,8 +5,8 @@ namespace GalaxusIntegration.Application.Commands.Products
     public class UpdateProductPriceCommand : IRequest<bool>
     {
         public string ProviderKey { get; set; }
-        public decimal NewPrice { get; set; }
-        public UpdateProductPriceCommand(string providerKey, decimal newPrice)
+        public double NewPrice { get; set; }
+        public UpdateProductPriceCommand(string providerKey, double newPrice)
         {
             ProviderKey = providerKey;
             NewPrice = newPrice;
